# openapi.yaml
openapi: 3.0.3
info:
  title: Landed Cost Calculation API
  description: |
    API สำหรับคำนวณต้นทุนนำเข้าสินค้า (Landed Cost) จากจีนเข้าสหรัฐอเมริกา
    รวมภาษีศุลกากรหลายชั้น (MFN, Section 301, Section 232, IEEPA)
    พร้อมระบบ temporal versioning, exclusion tracking และ audit trail

    ## ภาษีที่รองรับ
    | ประเภท | คำอธิบาย | อัตราทั่วไป |
    |--------|----------|------------|
    | MFN | อัตราฐาน Most Favored Nation | 0-9% |
    | Section 301 | ภาษีเฉพาะจีน (2018+) | 25% |
    | Section 232 | ภาษีความมั่นคงแห่งชาติ (เหล็ก/อลูมิเนียม) | 25% |
    | IEEPA | ภาษีตอบโต้ reciprocal (2025+) | 14.5% |

    ## Stacking Logic
    - MFN เป็นฐานเสมอ
    - Section 301 และ IEEPA **เลือกอันที่สูงกว่าอันเดียว**
    - Section 232 และ IEEPA **เลือกอันที่สูงกว่าอันเดียว**
  version: 1.0.0
  contact:
    name: API Support
    email: support@landed-cost.dev
  license:
    name: MIT
servers:
  - url: https://landed-cost.fly.dev
    description: Production
  - url: https://landed-cost-staging.fly.dev
    description: Staging
  - url: http://localhost:8000
    description: Local Development
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /hts-codes:
    get:
      tags: [HTS Codes]
      summary: ค้นหารหัส HTS
      description: |
        ค้นหารหัสพิกัดศุลกากร (HTS) ตาม prefix ของรหัส หรือ keyword ในคำอธิบาย
        รองรับ filter ตามระดับ: chapter (2 หลัก), heading (4 หลัก), subheading (6+ หลัก)
      operationId: searchHTSCodes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: q
          in: query
          description: ค้นหาตาม prefix ของรหัส (เช่น '8483') หรือ keyword (เช่น 'gear')
          required: false
          schema:
            type: string
            example: "8483"
        - name: level
          in: query
          description: กรองตามระดับลำดับชั้น
          required: false
          schema:
            type: string
            enum: [chapter, heading, subheading]
      responses:
        '200':
          description: รายการ HTS codes ที่ตรงตามเงื่อนไข
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/HTSCodeResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: "a1b2c3d4-5678-90ab-cdef-111111111111"
                    code: "8483.40"
                    level: subheading
                    description: "Gears and gearing; ball or roller screws; gear boxes and other speed changers"
                    parent_code: "8483"
                  - id: "a1b2c3d4-5678-90ab-cdef-222222222222"
                    code: "8483.50"
                    level: subheading
                    description: "Flywheels and pulleys, including pulley blocks"
                    parent_code: "8483"
                meta:
                  page: 1
                  page_size: 20
                  total_items: 2
                  total_pages: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [HTS Codes]
      summary: เพิ่มรหัส HTS ใหม่
      description: |
        เพิ่มรหัส HTS ใหม่เข้าระบบ ระบบจะกำหนด level อัตโนมัติจากรูปแบบรหัส:
        - 2 หลัก → chapter
        - 4 หลัก → heading
        - 6+ หลัก → subheading
      operationId: createHTSCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HTSCodeCreate'
            example:
              code: "8483.90"
              description: "Toothed wheels, chain sprockets and other transmission elements; parts"
              parent_code: "8483"
      responses:
        '201':
          description: สร้าง HTS code สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/HTSCodeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: รหัส HTS นี้มีอยู่ในระบบแล้ว
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: CONFLICT
                  message: "HTS code '8483.90' already exists"
                  details: null
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hts-codes/{code}:
    get:
      tags: [HTS Codes]
      summary: ดูรายละเอียด HTS code เฉพาะตัว
      description: ดึงข้อมูลรหัส HTS ที่ระบุ คืน 404 ถ้าไม่พบ
      operationId: getHTSCode
      parameters:
        - name: code
          in: path
          required: true
          description: รหัส HTS (เช่น '8483.40')
          schema:
            type: string
            example: "8483.40"
      responses:
        '200':
          description: พบข้อมูล HTS code
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/HTSCodeResponse'
              example:
                data:
                  id: "a1b2c3d4-5678-90ab-cdef-111111111111"
                  code: "8483.40"
                  level: subheading
                  description: "Gears and gearing; ball or roller screws; gear boxes and other speed changers"
                  parent_code: "8483"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /fx-rates:
    get:
      tags: [Exchange Rates]
      summary: ดูอัตราแลกเปลี่ยนย้อนหลัง
      description: |
        แสดงอัตราแลกเปลี่ยนสำหรับคู่สกุลเงินที่ระบุ
        ระบบใช้อัตรา ณ วันที่นำเข้าสำหรับคำนวณ
        ถ้าไม่มีอัตราตรงวัน จะ fallback ไปวันที่ใกล้ที่สุดก่อนหน้า
      operationId: listFXRates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: from_currency
          in: query
          description: สกุลเงินต้นทาง (ISO 4217)
          schema:
            type: string
            default: CNY
            example: CNY
        - name: to_currency
          in: query
          description: สกุลเงินปลายทาง (ISO 4217)
          schema:
            type: string
            default: USD
            example: USD
        - name: date_from
          in: query
          description: เริ่มต้นช่วงวันที่
          required: false
          schema:
            type: string
            format: date
            example: "2026-07-01"
        - name: date_to
          in: query
          description: สิ้นสุดช่วงวันที่
          required: false
          schema:
            type: string
            format: date
            example: "2026-08-01"
      responses:
        '200':
          description: รายการอัตราแลกเปลี่ยน
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FXRateResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: "fx-001"
                    from_currency: CNY
                    to_currency: USD
                    rate: 0.1389
                    date: "2026-08-01"
                  - id: "fx-002"
                    from_currency: CNY
                    to_currency: USD
                    rate: 0.1388
                    date: "2026-07-31"
                meta:
                  page: 1
                  page_size: 20
                  total_items: 2
                  total_pages: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Exchange Rates]
      summary: เพิ่มหรืออัปเดตอัตราแลกเปลี่ยน
      description: |
        เพิ่มอัตราแลกเปลี่ยนสำหรับวันที่ระบุ
        ถ้ามีอัตราของคู่สกุลเงิน+วันที่นี้อยู่แล้ว จะอัปเดตแทน (upsert)

        **รูปแบบ rate:** 1 หน่วย from_currency = rate หน่วย to_currency
        เช่น CNY→USD rate 0.1389 หมายถึง 1 หยวน = 0.1389 ดอลลาร์
      operationId: createFXRate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FXRateCreate'
            example:
              from_currency: CNY
              to_currency: USD
              rate: 0.1389
              date: "2026-08-01"
      responses:
        '201':
          description: สร้าง/อัปเดตอัตราแลกเปลี่ยนสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/FXRateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /calculations:
    get:
      tags: [Audit Trail]
      summary: ดูประวัติการคำนวณย้อนหลัง
      description: |
        แสดงรายการการคำนวณทั้งหมดที่เคยทำ เรียงตามเวลาล่าสุดก่อน
        ใช้สำหรับ audit — ดูว่าใครคำนวณอะไร เมื่อไร ใช้กฎเวอร์ชันไหน
      operationId: listCalculations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: hts_code
          in: query
          description: กรองตาม HTS code
          required: false
          schema:
            type: string
            example: "8483.40"
      responses:
        '200':
          description: รายการประวัติการคำนวณ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CalculationLogResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /calculations/{calculation_id}:
    get:
      tags: [Audit Trail]
      summary: ดูรายละเอียดการคำนวณครั้งที่ระบุ
      description: |
        ดึงข้อมูลการคำนวณครั้งเดียวแบบเต็ม ประกอบด้วย:
        - Input ที่ใช้
        - กฎภาษีที่ match
        - Exclusion ที่ใช้
        - ผลลัพธ์ทั้งหมด
      operationId: getCalculation
      parameters:
        - name: calculation_id
          in: path
          required: true
          description: UUID ของการคำนวณ
          schema:
            type: string
            format: uuid
            example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      responses:
        '200':
          description: รายละเอียดการคำนวณ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CalculationLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key สำหรับเข้าถึงระบบ ขอได้จากผู้ดูแลระบบ
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token สำหรับ production (future)
  parameters:
    PageParam:
      name: page
      in: query
      description: หมายเลขหน้า (เริ่มจาก 1)
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PageSizeParam:
      name: page_size
      in: query
      description: จำนวนรายการต่อหน้า
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
  responses:
...
    Unauthorized:
      description: ไม่มี API Key หรือ API Key ไม่ถูกต้อง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: "Missing API key. Provide it via the X-API-Key header."
              details: null
...
    NotFound:
      description: ไม่พบ resource ที่ร้องขอ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: NOT_FOUND
              message: "Resource not found"
              details: null
    ValidationError:
      description: ข้อมูลไม่ผ่าน validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: VALIDATION_ERROR
              message: "Request validation failed"
              details:
                - field: rate
                  message: "ensure this value is greater than or equal to 0"
    InternalServerError:
      description: เกิดข้อผิดพลาดภายในระบบ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: INTERNAL_ERROR
              message: "An unexpected error occurred"
              details: null
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: รหัส error มาตรฐาน
              example: BAD_REQUEST
            message:
              type: string
              description: คำอธิบาย error สำหรับ developer
              example: "HTS code '9999.99' not found"
            details:
              description: รายละเอียดเพิ่มเติม (validation errors, etc.)
              nullable: true
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 42
        total_pages:
          type: integer
          example: 3
...
    HTSCodeCreate:
      type: object
      required: [code, description]
      properties:
        code:
          type: string
          description: รหัส HTS (เช่น '84', '8483', '8483.40')
          minLength: 2
          maxLength: 20
          example: "8483.90"
        description:
          type: string
          description: คำอธิบายของรหัส HTS
          minLength: 3
          maxLength: 500
          example: "Toothed wheels, chain sprockets and other transmission elements"
        parent_code:
          type: string
          nullable: true
          description: รหัส HTS ของ parent (สำหรับลำดับชั้น)
          example: "8483"
    HTSCodeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: "8483.40"
        level:
          type: string
          enum: [chapter, heading, subheading]
          description: ระดับในลำดับชั้น
        description:
          type: string
          example: "Gears and gearing; gear boxes and other speed changers"
        parent_code:
          type: string
          nullable: true
          example: "8483"
...
    FXRateCreate:
      type: object
      required: [rate, date]
      properties:
        from_currency:
          type: string
          description: สกุลเงินต้นทาง (ISO 4217)
          minLength: 3
          maxLength: 3
          default: CNY
          example: CNY
        to_currency:
          type: string
          description: สกุลเงินปลายทาง (ISO 4217)
          minLength: 3
          maxLength: 3
          default: USD
          example: USD
        rate:
          type: number
          format: double
          description: อัตราแลกเปลี่ยน (1 หน่วย from = rate หน่วย to)
          minimum: 0
          exclusiveMinimum: true
          example: 0.1389
        date:
          type: string
          format: date
          description: วันที่ของอัตรานี้
          example: "2026-08-01"
    FXRateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from_currency:
          type: string
          example: CNY
        to_currency:
          type: string
          example: USD
        rate:
          type: number
          format: double
          example: 0.1389
        date:
          type: string
          format: date
          example: "2026-08-01"
    CalculationLogResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hts_code:
          type: string
          example: "8483.40"
        import_date:
          type: string
          format: date
        origin_country:
          type: string
          example: CN
        input_data:
          type: object
          description: Input ที่ใช้ในการคำนวณ (snapshot)
        matched_rules:
          type: array
          items:
            type: string
          description: รายการ rule IDs ที่ match
        exclusions_applied:
          type: array
          items:
            type: string
          description: รายการ exclusion IDs ที่ใช้
        result_data:
          type: object
          description: ผลลัพธ์การคำนวณ (snapshot)
        calculated_at:
          type: string
          format: date-time
          description: เวลาที่คำนวณ
          example: "2026-08-01T10:30:00Z"